use fuzzy matches to answer if the game is playable
